producer/requirements.txt
Flask
kafka-python
Werkzeug
